+++ Codive/Shared/Presentation/Components/CustomMultiSelectButton.swift
@@ -55,7 +55,7 @@ struct CustomMultiSelectButton: View {
             }
                         
             // Selection Buttons
-            FlowLayout(spacing: 8) {
+            CustomFlowLayout(spacing: 8) {
                 ForEach(options, id: \.self) { option in
                     SelectionButton(
                         title: option,
@@ -104,57 +104,24 @@ private struct SelectionButton: View {
     }
 }
 
-// MARK: - FlowLayout
-struct FlowLayout: Layout {
-    var spacing: CGFloat = 8
-    
-    func sizeThatFits(proposal: ProposedViewSize, subviews: Subviews, cache: inout ()) -> CGSize {
-        let result = FlowResult(
-            in: proposal.replacingUnspecifiedDimensions().width,
-            subviews: subviews,
-            spacing: spacing
+// MARK: - Preview
+#Preview {
+    VStack(spacing: 30) {
+        CustomMultiSelectButton(
+            title: "계절",
+            options: ["봄", "여름", "가을", "겨울"],
+            selectedOptions: .constant(["봄", "여름"]),
+            maxSelection: 4,
+            showRequiredMark: true
         )
-        return result.size
-    }
-    
-    func placeSubviews(in bounds: CGRect, proposal: ProposedViewSize, subviews: Subviews, cache: inout ()) {
-        let result = FlowResult(
-            in: bounds.width,
-            subviews: subviews,
-            spacing: spacing
-        )
-        for (index, subview) in subviews.enumerated() {
-            let position = CGPoint(
-                x: bounds.origin.x + result.positions[index].x,
-                y: bounds.origin.y + result.positions[index].y
-            )
-            subview.place(at: position, proposal: .unspecified)
-        }
-    }
-    struct FlowResult {
-        var size: CGSize = .zero
-        var positions: [CGPoint] = []
         
-        init(in maxWidth: CGFloat, subviews: Subviews, spacing: CGFloat) {
-            var x: CGFloat = 0
-            var y: CGFloat = 0
-            var lineHeight: CGFloat = 0
-            
-            for subview in subviews {
-                let size = subview.sizeThatFits(.unspecified)
-                
-                if x + size.width > maxWidth && x > 0 {
-                    x = 0
-                    y += lineHeight + spacing
-                    lineHeight = 0
-                }
-                
-                positions.append(CGPoint(x: x, y: y))
-                lineHeight = max(lineHeight, size.height)
-                x += size.width + spacing
-            }
-            
-            self.size = CGSize(width: maxWidth, height: y + lineHeight)
-        }
+        CustomMultiSelectButton(
+            title: "스타일",
+            options: ["캐주얼", "포멀", "스트릿", "빈티지", "미니멀"],
+            selectedOptions: .constant(["캐주얼"]),
+            maxSelection: 3,
+            showRequiredMark: false
+        )
     }
+    .padding(20)
 }
